proxy: simplify Codex desktop forwarding
